name wall working with asynchronicity, still needs tests
