+++ src/Win.js
@@ -3,7 +3,7 @@ import React, { Component } from 'react'
 export default class Win extends Component {
   render () {
     return (
-      <div className='WIN'>
+      <div className='win'>
         <div id='gameStatus' >
           <p>Days on the trail: {this.props.days}</p>
           <p>Miles: {this.props.miles}</p>
